requirements.txt
pandas==2.2.3
requests==2.32.3
openai==1.51.2
python-dotenv==1.0.0
